Update is Power of 2.cpp

+++ Number Theory/is Power of 2.cpp
@@ -1,4 +1,4 @@
-bool is_pow_of_2(ll x){
+bool is_pow_of_2(int x){
     if(x == 0) return false;
     if(x & (x - 1)) return false;
     else return true;
